+++ packages/e2e-tests/src/tests/dust.remote.test.ts
@@ -13,7 +13,7 @@
 import { describe, test, expect } from 'vitest';
 import * as rx from 'rxjs';
 import { useTestContainersFixture } from './test-fixture.js';
-import * as ledger from '@midnight-ntwrk/ledger-v6';
+import * as ledger from '@midnight-ntwrk/ledger-v7';
 import * as utils from './utils.js';
 import { logger } from './logger.js';
 import { WalletFacade } from '@midnight-ntwrk/wallet-sdk-facade';
@@ -35,7 +35,7 @@ describe('Dust tests', () => {
   let walletFacade: WalletFacade;
   let walletKeystore: UnshieldedKeystore;
 
-  beforeAll(async () => {
+  beforeEach(async () => {
     const fixture = getFixture();
     const networkId = fixture.getNetworkId();
     walletFacade = utils.buildWalletFacade(seed, fixture);
@@ -44,9 +44,10 @@ describe('Dust tests', () => {
     await walletFacade.start(walletSecretKey, walletDustSecretKey);
   });
 
-  afterAll(async () => {
-    await utils.closeWallet(walletFacade);
-  }, 20_000);
+  afterEach(async () => {
+    await walletFacade.stop();
+    logger.info('Wallet stopped');
+  });
 
   test(
     'Able to register Night tokens for Dust generation @healthcheck',
@@ -121,9 +122,9 @@ describe('Dust tests', () => {
       logger.info(`Registered night UTXOs: ${inspect(registeredNightUtxos, { depth: null })}`);
       expect(registeredNightUtxos.length).toBeGreaterThan(0);
 
-      const deregisterTokens = 2;
+      const firstRegisteredNightUtxo = registeredNightUtxos[0];
       const dustDeregistrationRecipe = await walletFacade.deregisterFromDustGeneration(
-        registeredNightUtxos.slice(0, deregisterTokens),
+        [firstRegisteredNightUtxo],
         walletKeystore.getPublicKey(),
         (payload) => walletKeystore.signData(payload),
       );
@@ -141,11 +142,8 @@ describe('Dust tests', () => {
 
       const finalizedDustTx = await walletFacade.finalizeTransaction(balancedTransactionRecipe);
       const dustDeregistrationTxid = await walletFacade.submitTransaction(finalizedDustTx);
+      expect(dustDeregistrationTxid).toBeDefined();
       logger.info(`Dust de-registration tx id: ${dustDeregistrationTxid}`);
-      const finalState = await utils.waitForSyncFacade(walletFacade);
-      const finalDustBalance = finalState.dust.walletBalance(new Date());
-
-      expect(finalDustBalance).toBeLessThan(initialDustBalance);
     },
     timeout,
   );