+++ packages/unshielded-wallet/src/v1/test/testUtils.ts
@@ -10,8 +10,8 @@
 // WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
 // See the License for the specific language governing permissions and
 // limitations under the License.
-import * as ledger from '@midnight-ntwrk/ledger-v6';
-import { sampleIntentHash } from '@midnight-ntwrk/ledger-v6';
+import * as ledger from '@midnight-ntwrk/ledger-v7';
+import { sampleIntentHash } from '@midnight-ntwrk/ledger-v7';
 import { UnshieldedUpdate, UpdateStatus, UtxoWithMeta } from '../UnshieldedState.js';
 
 export const generateMockUtxoWithMeta = (owner: string, type: string): UtxoWithMeta =>