+++ packages/facade/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @midnight-ntwrk/wallet-sdk-facade
 
+## 1.0.0-beta.13
+
+### Patch Changes
+
+- Updated dependencies [aef8d4b]
+  - @midnight-ntwrk/wallet-sdk-unshielded-wallet@1.0.0-beta.15
+  - @midnight-ntwrk/wallet-sdk-shielded@1.0.0-beta.13
+  - @midnight-ntwrk/wallet-sdk-dust-wallet@1.0.0-beta.12
+
 ## 1.0.0-beta.12
 
 ### Patch Changes